25-05-23 jae 글쓰기 기능

+++ SEMIProject-boot/src/main/java/edu/kh/semi/board/controller/FreeBoardController.java
@@ -53,7 +53,7 @@ public class FreeBoardController {
 	public String Write(Board board, HttpSession session,
 			@RequestParam(value = "boardImage", required = false) MultipartFile image) {
 		Member loginMember = (Member) session.getAttribute("loginMember");
-		board.setMemberNo((long) loginMember.getMemberNo());
+		board.setMemberNo((int) loginMember.getMemberNo());
 		board.setBoardCode(2);
 
 		int result = service.insertFreeBoard(board);
